python_backup/requirements.txt
streamlit
yfinance
pandas
notion-client
python-dotenv
google-genai
plotly
